Bigdata/requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow-cpu
beautifulsoup4
altair
Pillow
requests
joblib
matplotlib